Read compiler configuration from settings.json first.
The settings map a compiler name (the one passed to dub with
--compiler= option) to a configuration that includes the path
to the compiler binary, dflags (e.g. architecture, abi, cpu,
particular cpu features), and lflags (e.g. paths to libraries
specific to the target platform).

If a compiler with the specific name is not found in settings.json,
the previous behavior is used.

The intention of this change is to make it possible or at least
more convenient to use dub for cross-compilation, which usually
implies the need to pass many additional details about the target
platform to the compiler and linker. It isolates dub from this
complexity, yet it should allow the user to configure any
arbitrary toolchain and enable it in dub with a single command
line switch.